requirements.txt
appdirs==1.4.3
asgiref==3.3.4
async-timeout==3.0.1
attrs==21.2.0
boto3==1.17.84
botocore==1.20.84
CacheControl==0.12.6
certifi==2019.11.28
# required latest version of cffi
cffi
chardet==3.0.4
click==7.1.2
colorama==0.4.3
contextlib2==0.6.0
cryptography==3.4.7
distlib==0.3.0
distro==1.4.0
dnspython==2.1.0
gunicorn==20.1.0
h11==0.12.0
html5lib==1.0.1
httptools==0.2.0
idna==2.8
ipaddr==2.2.0
jmespath==0.10.0
#jwt==1.2.0
lockfile==0.12.2
msgpack==0.6.2
multidict==5.1.0
packaging==20.3
pep517==0.8.2
progress==1.5
pycparser==2.20
#pyJWT==1.7.1
pyJWT==2.6.0
pymongo==3.11.4
pyparsing==2.4.6
python-dateutil==2.8.1
python-multipart==0.0.5
pytoml==0.1.21
requests==2.22.0
retrying==1.3.3
s3transfer==0.4.2
six==1.14.0
starlette==0.14.2
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.25.8
uvicorn==0.13.4
uvloop==0.15.2
webencodings==0.5.1
websockets==9.1
yarl==1.6.3
